added high resolution scrolling
This still requires fixing to make it work correctly.

+++ cores/nRF5/Adafruit_TinyUSB_Core/tinyusb/src/class/hid/hid.h
@@ -155,7 +155,7 @@ typedef struct ATTR_PACKED
   uint8_t buttons; /**< buttons mask for currently pressed buttons in the mouse. */
   int8_t  x;       /**< Current delta x movement of the mouse. */
   int8_t  y;       /**< Current delta y movement on the mouse. */
-  int8_t  wheel;   /**< Current delta wheel movement on the mouse. */
+  int16_t  wheel;   /**< Current delta wheel movement on the mouse. */
   int8_t  pan;     // using AC Pan
 } hid_mouse_report_t;
